Limpeza e organização de codigo
