--- analisador-lexico-uni1/Resultado-saida_backup.txt
@@ -1,90 +1,0 @@
-
-TOTAL DE CLASSES: 236
-TOTAL DE INDIVIDUOS: 0
-TOTAL DE PROPRIEDADES: 173
-TOTAL DE TIPOS DE DADOS: 0
-TOTAL DE PALAVRAS RESERVADAS: 253
-TOTAL DE NAMESPACE: 0
-TOTAL DE CARACTERES ESPECIAIS: 203
-TOTAL DE CARDINALIDADES: 54
-
-
-
-['Class:' : 'PALAVRA_RESERVADA', quantidade: 41]
-['Activity' : 'CLASSE', quantidade: 5]
-['SubClassOf:' : 'PALAVRA_RESERVADA', quantidade: 35]
-['FunctionalComplex' : 'CLASSE', quantidade: 4]
-[',' : 'CARACTERE_ESPECIAL', quantidade: 95]
-['participatedIn' : 'PROPRIEDADE', quantidade: 17]
-['some' : 'PALAVRA_RESERVADA', quantidade: 58]
-['DataExchange' : 'CLASSE', quantidade: 4]
-['Actor' : 'CLASSE', quantidade: 4]
-['BrokerServiceProvider' : 'CLASSE', quantidade: 4]
-['IntermediaryParticipant' : 'CLASSE', quantidade: 4]
-['mediates' : 'PROPRIEDADE', quantidade: 51]
-['MetadataAssignment' : 'CLASSE', quantidade: 6]
-['MetadataRetrival' : 'CLASSE', quantidade: 6]
-['Certificate' : 'CLASSE', quantidade: 4]
-['Resource' : 'CLASSE', quantidade: 10]
-['CertificationAct' : 'CLASSE', quantidade: 5]
-['Certification' : 'CLASSE', quantidade: 2]
-['historicallyDependsOn' : 'PROPRIEDADE', quantidade: 20]
-['only' : 'PROPRIEDADE', quantidade: 29]
-['exactly' : 'PALAVRA_RESERVADA', quantidade: 18]
-['1' : 'CARDINALIDADE', quantidade: 53]
-['EquivalentTo:' : 'PALAVRA_RESERVADA', quantidade: 7]
-['Relator' : 'CLASSE', quantidade: 9]
-['and' : 'PALAVRA_RESERVADA', quantidade: 46]
-['(' : 'CARACTERE_ESPECIAL', quantidade: 54]
-['EvaluationReport' : 'CLASSE', quantidade: 6]
-[')' : 'CARACTERE_ESPECIAL', quantidade: 54]
-['or' : 'PALAVRA_RESERVADA', quantidade: 10]
-['CertificationBody' : 'CLASSE', quantidade: 3]
-['Evaluated' : 'CLASSE', quantidade: 5]
-['min' : 'PALAVRA_RESERVADA', quantidade: 36]
-['emitsCertificate' : 'PROPRIEDADE', quantidade: 1]
-['ConditionalClaim' : 'CLASSE', quantidade: 6]
-['IntrinsicMode' : 'CLASSE', quantidade: 4]
-['inheresIn' : 'PROPRIEDADE', quantidade: 15]
-['DataCustomer' : 'CLASSE', quantidade: 17]
-['externallyDependsOn' : 'PROPRIEDADE', quantidade: 12]
-['DataSupplier' : 'CLASSE', quantidade: 19]
-['ConditionalCommitment' : 'CLASSE', quantidade: 6]
-['Connector' : 'CLASSE', quantidade: 3]
-['InvokeDataOperation' : 'CLASSE', quantidade: 7]
-['enforces' : 'PROPRIEDADE', quantidade: 1]
-['DataUsagePolicy' : 'CLASSE', quantidade: 6]
-['CoreParticipant' : 'CLASSE', quantidade: 4]
-['Counterpart' : 'CLASSE', quantidade: 3]
-['OffereeUnconditionalAgreement' : 'CLASSE', quantidade: 6]
-['CounterpartContribution' : 'CLASSE', quantidade: 4]
-['Event' : 'CLASSE', quantidade: 4]
-['CounterpartContributionType' : 'CLASSE', quantidade: 3]
-['ConcreteIndividualType' : 'CLASSE', quantidade: 2]
-['instantiate' : 'PROPRIEDADE', quantidade: 4]
-['Data' : 'CLASSE', quantidade: 7]
-['Metadata' : 'CLASSE', quantidade: 4]
-['inverse' : 'PROPRIEDADE', quantidade: 3]
-['DataSovereigntyAct' : 'CLASSE', quantidade: 4]
-['Ownership' : 'CLASSE', quantidade: 3]
-['OfferorUnconditionalAgreement' : 'CLASSE', quantidade: 6]
-['Offer' : 'CLASSE', quantidade: 5]
-['calls' : 'PROPRIEDADE', quantidade: 1]
-['DataOperation' : 'CLASSE', quantidade: 4]
-['EconomicOffering' : 'CLASSE', quantidade: 7]
-['DisjointWith:' : 'PALAVRA_RESERVADA', quantidade: 2]
-['2' : 'CARDINALIDADE', quantidade: 1]
-['composedBy' : 'PROPRIEDADE', quantidade: 10]
-['OfferedContribution' : 'CLASSE', quantidade: 4]
-['creates' : 'PROPRIEDADE', quantidade: 1]
-['owns' : 'PROPRIEDADE', quantidade: 1]
-['responds' : 'PROPRIEDADE', quantidade: 1]
-['DigitalAsset' : 'CLASSE', quantidade: 3]
-['EconomicAgreement' : 'CLASSE', quantidade: 1]
-['wasCreatedIn' : 'PROPRIEDADE', quantidade: 4]
-['Evaluation' : 'CLASSE', quantidade: 2]
-['EvaluationAct' : 'CLASSE', quantidade: 5]
-['EvaluationFacility' : 'CLASSE', quantidade: 3]
-['emitsReport' : 'PROPRIEDADE', quantidade: 1]
-['describes' : 'PROPRIEDADE', quantidade: 1]
-['OfferedContributionType' : 'CLASSE', quantidade: 3]
