--- analisador-lexico-uni1/codigo2.txt
@@ -1,102 +1,0 @@
-Class: CounterpartContribution
-
-        
-    SubClassOf: 
-        Event,
-        participatedIn only DataCustomer,
-        participatedIn min 1 DataCustomer
-    
-    
-Class: CounterpartContributionType
-
-        
-    SubClassOf: 
-        ConcreteIndividualType,
-        instantiate some CounterpartContribution,
-        instantiate only CounterpartContribution
-    
-    
-Class: Data
-
-        
-    SubClassOf: 
-        Resource,
-        inheresIn some Metadata,
-        inverse mediates some DataSovereigntyAct,
-        inverse mediates some InvokeDataOperation,
-        inverse mediates exactly 1 Ownership
-    
-    
-Class: DataCustomer
-
-        
-    SubClassOf: 
-        CoreParticipant,
-        externallyDependsOn some ConditionalCommitment,
-        externallyDependsOn some OfferorUnconditionalAgreement,
-        participatedIn some Offer,
-        calls some DataOperation,
-        inheresIn some ConditionalClaim,
-        inheresIn some OffereeUnconditionalAgreement,
-        mediates some EconomicOffering,
-        mediates some InvokeDataOperation,
-        mediates some MetadataRetrival
-    
-    DisjointWith: 
-        DataSupplier
-    
-    
-Class: DataExchange
-
-        
-    EquivalentTo: 
-        Event
-         and (participatedIn some Activity)
-         and (participatedIn some Resource)
-         and (participatedIn min 1 DataCustomer)
-         and (participatedIn min 1 DataSupplier)
-         and (participatedIn min 2 Actor)
-         and (composedBy exactly 1 CounterpartContribution)
-         and (composedBy exactly 1 OfferedContribution)
-    
-    
-Class: DataOperation
-
-       
-    SubClassOf: 
-        Activity,
-         mediates some InvokeDataOperation,
-         mediates only InvokeDataOperation
-    
-    
-Class: DataSovereigntyAct
-
-        
-    SubClassOf: 
-        Relator,
-        mediates some Data,
-        mediates some DataUsagePolicy,
-        mediates min 1 DataSupplier
-    
-    
-Class: DataSupplier
-
-        
-    SubClassOf: 
-        CoreParticipant,
-        externallyDependsOn some ConditionalClaim,
-        externallyDependsOn some OffereeUnconditionalAgreement,
-        participatedIn some Offer,
-        creates some DataUsagePolicy,
-        owns some Data,
-        responds some DataOperation,
-        inheresIn some ConditionalCommitment,
-        inheresIn some OfferorUnconditionalAgreement,
-        mediates some DataSovereigntyAct,
-        mediates some EconomicOffering,
-        mediates some MetadataAssignment,
-        mediates some Ownership
-    
-    DisjointWith: 
-        DataCustomer
-
